feat(core): upload conversion bundles
metriport/metriport-internal#1040
