+++ packages/core/src/external/surescripts/command/convert-batch-response/convert-batch-response-direct.ts
@@ -1,6 +1,4 @@
 import { MetriportError } from "@metriport/shared";
-import { Config } from "../../../../util/config";
-import { S3Utils } from "@metriport/core/external/aws/s3";
 import { SurescriptsConvertBatchResponseHandler } from "./convert-batch-response";
 import { SurescriptsReplica } from "../../replica";
 import {
@@ -8,8 +6,7 @@ import {
   SurescriptsFileIdentifier,
   SurescriptsRequester,
 } from "../../types";
-import { convertBatchResponseToFhirBundles } from "../../fhir-converter";
-import { makeConversionBundleFileName } from "../../file/file-names";
+import { convertBatchResponseToFhirBundles, uploadConversionBundle } from "../../fhir-converter";
 
 export class SurescriptsConvertBatchResponseHandlerDirect
   implements SurescriptsConvertBatchResponseHandler
@@ -36,15 +33,9 @@ export class SurescriptsConvertBatchResponseHandlerDirect
       );
     }
 
-    const conversionBucket = new S3Utils(Config.getAWSRegion());
     const conversionBundles = await convertBatchResponseToFhirBundles(responseFileContent);
     for (const { patientId, bundle } of conversionBundles) {
-      const fileName = makeConversionBundleFileName(cxId, patientId);
-      await conversionBucket.uploadFile({
-        bucket: Config.getPharmacyConversionBucketName(),
-        key: fileName,
-        file: Buffer.from(JSON.stringify(bundle)),
-      });
+      await uploadConversionBundle({ bundle, cxId, patientId });
     }
     return conversionBundles;
   }
